+++ CHANGELOG.md
@@ -34,6 +34,7 @@
 - Enh #46: Add `InputFile` class for HTML `<input type="file">` element with attributes and rendering capabilities (@terabytesoftw)
 - Bug #47: Update PHPDoc in `tests` and add new tests for `on*` attribute (@terabytesoftw)
 - Enh #48: Add `InputEmail` class for HTML `<input type="email">` element with attributes and rendering capabilities (@terabytesoftw)
+- Enh #49: Add `InputDateTimeLocal` class for HTML `<input type="datetime-local">` element with attributes and rendering capabilities (@terabytesoftw)
 
 ## 0.3.0 March 31, 2024